+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Released: Upcoming
 - Added pagination to the list view
 - Added the ability to hand in default values to the create form
 - Added edit and delete buttons to the children table of the entities show page
+- CRUDMySQLData::listEntries handles null filter values now correctly
 
 ## 0.9.5
 Released: 2014-11-07